Output benchmark to php file

+++ run.php
@@ -8,18 +8,30 @@ $dimensions = array(
     array(200, 150),
     array(150, 150)
 );
+$benchmark = array();
 
 while ($file = readdir($dir)) {
     if (is_dir($file) && $file[0]!='.') {
-        echo "* Processing $file...\n";
-        foreach ($dimensions as $dimension) {
-            list($w, $h) = $dimension;
-            $t = microtime(true);
-            `rm $file/plate_*.stl 2> /dev/null`;
-            `plater -t 4 -W $w -H $h -s 1.75 $file/plater.conf 2> /dev/null`;
-            $plates = count(explode("\n",`ls $file/plate_*.stl 2> /dev/null`))-1;
-            $e = round(microtime(true)-$t,2);
-            echo "~> Width: $w, Height: $h, Time: $e, Plates: $plates\n";
+        if (file_exists($file.'/plater.conf')) {
+            echo "* Processing $file...\n";
+            $benchmark[$file] = array();
+            foreach ($dimensions as $dimension) {
+                list($w, $h) = $dimension;
+                $t = microtime(true);
+                `rm $file/plate_*.stl 2> /dev/null`;
+                `plater -t 4 -W $w -H $h -s 1.75 $file/plater.conf 2> /dev/null`;
+                $plates = count(explode("\n",`ls $file/plate_*.stl 2> /dev/null`))-1;
+                $e = round(microtime(true)-$t,2);
+                $benchmark[$file][] = array(
+                    'width' => $w,
+                    'height' => $h,
+                    'time' => $e,
+                    'plates' => $plates
+                );
+                echo "~> Width: $w, Height: $h, Time: $e, Plates: $plates\n";
+            }
         }
     }
 }
+
+fwrite(STDERR, var_export($benchmark, true));
